serialization test renamed
