--- zenoh-ts/tests/src/z_serialization_organized.ts
@@ -1,1361 +1,0 @@
-//
-// Copyright (c) 2025 ZettaScale Technology
-//
-// This program and the accompanying materials are made available under the
-// terms of the Eclipse Public License 2.0 which is available at
-// http://www.eclipse.org/legal/epl-2.0, or the Apache License, Version 2.0
-// which is available at https://www.apache.org/licenses/LICENSE-2.0.
-//
-// SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
-//
-// Contributors:
-//   ZettaScale Zenoh Team, <zenoh@zettascale.tech>
-//
-
-import { ZBytesSerializer, ZBytesDeserializer, ZSerializeable, ZDeserializeable, zserialize, zdeserialize, NumberFormat, BigIntFormat, ZS, ZD } from "@eclipse-zenoh/zenoh-ts/ext";
-import { assertEquals, assert } from "https://deno.land/std@0.192.0/testing/asserts.ts";
-
-// Test classes
-class CustomStruct implements ZSerializeable, ZDeserializeable {
-    constructor(
-        public vd: number[] = [],
-        public i: number = 0,
-        public s: string = ""
-    ) {}
-
-    serializeWithZSerializer(serializer: ZBytesSerializer): void {
-        serializer.serialize(this.vd, ZS.array(ZS.number()));
-        serializer.serialize(this.i, ZS.number(NumberFormat.Int32));
-        serializer.serialize(this.s, ZS.string());
-    }
-
-    deserializeWithZDeserializer(deserializer: ZBytesDeserializer): void {
-        this.vd = deserializer.deserialize(ZD.array(ZD.number()));
-        this.i = deserializer.deserialize(ZD.number(NumberFormat.Int32));
-        this.s = deserializer.deserialize(ZD.string());
-    }
-}
-
-class MixedTypedArrayCollection implements ZSerializeable, ZDeserializeable {
-    uint16: Uint16Array;
-    int32: Int32Array;
-    float64: Float64Array;
-    
-    constructor(
-        uint16: Uint16Array = new Uint16Array(0),
-        int32: Int32Array = new Int32Array(0),
-        float64: Float64Array = new Float64Array(0)
-    ) {
-        this.uint16 = uint16;
-        this.int32 = int32;
-        this.float64 = float64;
-    }
-    
-    serializeWithZSerializer(serializer: ZBytesSerializer): void {
-        serializer.serialize(this.uint16, ZS.uint16array());
-        serializer.serialize(this.int32, ZS.int32array());
-        serializer.serialize(this.float64, ZS.float64array());
-    }
-    
-    deserializeWithZDeserializer(deserializer: ZBytesDeserializer): void {
-        this.uint16 = deserializer.deserialize(ZD.uint16array());
-        this.int32 = deserializer.deserialize(ZD.int32array());
-        this.float64 = deserializer.deserialize(ZD.float64array());
-    }
-}
-
-// Helper functions
-function compareFloatArrays(a: ArrayLike<number>, b: ArrayLike<number>, tolerance = 0.0001): boolean {
-    if (a.length !== b.length) return false;
-    return Array.from(a).every((val, idx) => Math.abs(val - b[idx]) < tolerance);
-}
-
-/**
- * 1. PRIMITIVE TYPE TESTS
- */
-Deno.test("Serialization - Primitive Number Types", () => {
-    // === UINT FORMATS ===
-    // Test Uint8
-    for (const val of [0, 5, 127, 255]) {
-        const bytes = zserialize(val, ZS.number(NumberFormat.Uint8));
-        const result = zdeserialize(ZD.number(NumberFormat.Uint8), bytes);
-        assertEquals(result, val, `Uint8 serialization failed for value: ${val}`);
-    }
-
-    // Test Uint16
-    for (const val of [0, 500, 65535]) {
-        const bytes = zserialize(val, ZS.number(NumberFormat.Uint16));
-        const result = zdeserialize(ZD.number(NumberFormat.Uint16), bytes);
-        assertEquals(result, val, `Uint16 serialization failed for value: ${val}`);
-    }
-
-    // Test Uint32
-    for (const val of [0, 50000, 4294967295]) {
-        const bytes = zserialize(val, ZS.number(NumberFormat.Uint32));
-        const result = zdeserialize(ZD.number(NumberFormat.Uint32), bytes);
-        assertEquals(result, val, `Uint32 serialization failed for value: ${val}`);
-    }
-
-    // Test Uint64 (within safe integer range)
-    for (const val of [0, 1000000, Number.MAX_SAFE_INTEGER]) {
-        const bytes = zserialize(val, ZS.number(NumberFormat.Uint64));
-        const result = zdeserialize(ZD.number(NumberFormat.Uint64), bytes);
-        assertEquals(result, val, `Uint64 serialization failed for value: ${val}`);
-    }
-
-    // === INT FORMATS ===
-    // Test Int8
-    for (const val of [-128, -5, 0, 5, 127]) {
-        const bytes = zserialize(val, ZS.number(NumberFormat.Int8));
-        const result = zdeserialize(ZD.number(NumberFormat.Int8), bytes);
-        assertEquals(result, val, `Int8 serialization failed for value: ${val}`);
-    }
-
-    // Test Int16
-    for (const val of [-32768, -500, 0, 500, 32767]) {
-        const bytes = zserialize(val, ZS.number(NumberFormat.Int16));
-        const result = zdeserialize(ZD.number(NumberFormat.Int16), bytes);
-        assertEquals(result, val, `Int16 serialization failed for value: ${val}`);
-    }
-
-    // Test Int32
-    for (const val of [-2147483648, -50000, 0, 50000, 2147483647]) {
-        const bytes = zserialize(val, ZS.number(NumberFormat.Int32));
-        const result = zdeserialize(ZD.number(NumberFormat.Int32), bytes);
-        assertEquals(result, val, `Int32 serialization failed for value: ${val}`);
-    }
-
-    // Test Int64 (within safe integer range)
-    for (const val of [-Number.MAX_SAFE_INTEGER, -1000000, 0, 1000000, Number.MAX_SAFE_INTEGER]) {
-        const bytes = zserialize(val, ZS.number(NumberFormat.Int64));
-        const result = zdeserialize(ZD.number(NumberFormat.Int64), bytes);
-        assertEquals(result, val, `Int64 serialization failed for value: ${val}`);
-    }
-
-    // === FLOAT FORMATS ===
-    // Test Float32
-    for (const val of [-3.14, 0, 0.5, 3.14]) {
-        const bytes = zserialize(val, ZS.number(NumberFormat.Float32));
-        const result = zdeserialize(ZD.number(NumberFormat.Float32), bytes);
-        assert(Math.abs(val - result) < 0.0001, `Float32 serialization failed for value: ${val}`);
-    }
-
-    // Test Float64
-    for (const val of [-123.45, 0, 123.45]) {
-        const bytes = zserialize(val, ZS.number(NumberFormat.Float64));
-        const result = zdeserialize(ZD.number(NumberFormat.Float64), bytes);
-        assert(Math.abs(val - result) < 0.0001, `Float64 serialization failed for value: ${val}`);
-    }
-});
-
-Deno.test("Serialization - BigInt Types", () => {
-    // Test BigInt Uint64
-    for (const val of [0n, 500000000000n, 18446744073709551615n]) {
-        const bytes = zserialize(val, ZS.bigint(BigIntFormat.Uint64));
-        const result = zdeserialize(ZD.bigint(BigIntFormat.Uint64), bytes);
-        assertEquals(result, val, `BigInt Uint64 serialization failed for value: ${val}`);
-    }
-
-    // Test BigInt Int64
-    for (const val of [-9223372036854775808n, -500000000000n, 0n, 500000000000n, 9223372036854775807n]) {
-        const bytes = zserialize(val, ZS.bigint(BigIntFormat.Int64));
-        const result = zdeserialize(ZD.bigint(BigIntFormat.Int64), bytes);
-        assertEquals(result, val, `BigInt Int64 serialization failed for value: ${val}`);
-    }
-
-    // Test number larger than 32 bit but less than 53 bit
-    const large32BitNum = 0x1_0000_0001; // Just over 32 bits
-    const bytes = zserialize(large32BitNum, ZS.number(NumberFormat.Int64));
-    const result = zdeserialize(ZD.number(NumberFormat.Int64), bytes);
-    assertEquals(result, large32BitNum, "large number (>32bit) serialization failed");
-
-    // Test that deserializing number larger than 53 bit throws error
-    const tooLargeNum = 9876543210123456789n; // Larger than 53 bits and less than 64 bits
-    assert(tooLargeNum > BigInt(Number.MAX_SAFE_INTEGER), "number is not larger than 53 bits");
-    assert(tooLargeNum < BigInt(2) ** BigInt(64), "number is not less than 2^64");
-
-    let threw = false;
-    const largeBytes = zserialize(tooLargeNum, ZS.bigint(BigIntFormat.Uint64));
-    try {
-        zdeserialize(ZD.number(NumberFormat.Uint64), largeBytes);
-    } catch (e) {
-        threw = true;
-        assert(e instanceof Error && e.message.includes("exceeds the safe range"), 
-               "wrong error message for large number deserialization");
-    }
-    assert(threw, "Should throw when deserializing bigint > MAX_SAFE_INTEGER as number");
-});
-
-Deno.test("Serialization - Boolean", () => {
-    // Test boolean true
-    let bytes = zserialize(true, ZS.boolean());
-    let result = zdeserialize(ZD.boolean(), bytes);
-    assertEquals(result, true, "Boolean true serialization failed");
-
-    // Test boolean false
-    bytes = zserialize(false, ZS.boolean());
-    result = zdeserialize(ZD.boolean(), bytes);
-    assertEquals(result, false, "Boolean false serialization failed");
-});
-
-/**
- * 2. CONTAINER TYPES TESTS
- */
-Deno.test("Serialization - String", () => {
-    // Test empty string
-    let str = "";
-    let bytes = zserialize(str, ZS.string());
-    let result = zdeserialize(ZD.string(), bytes);
-    assertEquals(result, str, "Empty string serialization failed");
-
-    // Test simple string
-    str = "abcdefg";
-    bytes = zserialize(str);
-    result = zdeserialize(ZD.string(), bytes);
-    assertEquals(result, str, "String serialization failed");
-
-    // Test string with special characters
-    str = "Special: áéíóú ñ 你好";
-    bytes = zserialize(str);
-    result = zdeserialize(ZD.string(), bytes);
-    assertEquals(result, str, "String with special characters serialization failed");
-});
-
-Deno.test("Serialization - Arrays", () => {
-    // Test empty array
-    const emptyArray: number[] = [];
-    let bytes = zserialize(emptyArray, ZS.array(ZS.number()));
-    let result = zdeserialize(ZD.array(ZD.number()), bytes);
-    assertEquals(result, emptyArray, "Empty array serialization failed");
-    
-    // Test homogeneous number array
-    const numArray = [0.5, 1.0, 2.0];
-    bytes = zserialize(numArray, ZS.array(ZS.number()));
-    result = zdeserialize(ZD.array(ZD.number()), bytes);
-    assertEquals(result, numArray, "Number array serialization failed");
-
-    // Test homogeneous string array
-    const strArray = ["abc", "def", "hij"];
-    bytes = zserialize(strArray, ZS.array(ZS.string()));
-    const strResult = zdeserialize(ZD.array(ZD.string()), bytes);
-    assertEquals(strResult, strArray, "String array serialization failed");
-
-    // Test array of floats
-    const floats = [0.1, 0.2, -0.5, 1000.578];
-    bytes = zserialize(floats, ZS.array(ZS.number(NumberFormat.Float32)));
-    const floatResult = zdeserialize(ZD.array(ZD.number(NumberFormat.Float32)), bytes);
-    assert(floatResult.every((val, idx) => Math.abs(val - floats[idx]) < 0.0001), 
-           "Array of floats serialization failed");
-           
-    // Test nested array
-    const nestedArray = [[1, 2, 3], [4, 5, 6]];
-    bytes = zserialize(nestedArray, ZS.array(ZS.array(ZS.number())));
-    const nestedResult = zdeserialize(ZD.array(ZD.array(ZD.number())), bytes);
-    assertEquals(nestedResult, nestedArray, "Nested array serialization failed");
-
-    // Test array of bigints
-    const int64Array = [-100n, 500n, 100000n, -20000000n];
-    bytes = zserialize(int64Array, ZS.array(ZS.bigint()));
-    const int64Result = zdeserialize(ZD.array(ZD.bigint()), bytes);
-    assertEquals(int64Result, int64Array, "Array of bigints serialization failed");
-});
-
-Deno.test("Serialization - Sets and Maps", () => {
-    // Test Set conversion to/from array
-    const set = new Set([1, 2, 3, -5, 10000, -999999999]);
-    const setBytes = zserialize(Array.from(set));
-    const setResult = new Set(zdeserialize(ZD.array(ZD.number()), setBytes));
-    assertEquals(setResult, set, "Set serialization failed");
-
-    // Test empty Map
-    const emptyMap = new Map<number, string>();
-    let mapBytes = zserialize(emptyMap);
-    let mapResult = zdeserialize(ZD.map(ZD.number(), ZD.string()), mapBytes);
-    assertEquals(mapResult, emptyMap, "Empty map serialization failed");
-
-    // Test Map with entries
-    const map = new Map<number, string>();
-    map.set(100, "abc");
-    map.set(10000, "def");
-    map.set(2000000000, "hij");
-    mapBytes = zserialize(map);
-    mapResult = zdeserialize(ZD.map(ZD.number(), ZD.string()), mapBytes);
-    assertEquals(mapResult, map, "Map serialization failed");
-});
-
-Deno.test("Serialization - Custom Objects", () => {
-    // Test empty CustomStruct
-    const emptyStruct = new CustomStruct();
-    let bytes = zserialize(emptyStruct);
-    let result = zdeserialize(ZD.object(CustomStruct), bytes);
-    assertEquals(result.vd, emptyStruct.vd, "Empty CustomStruct.vd serialization failed");
-    assertEquals(result.i, emptyStruct.i, "Empty CustomStruct.i serialization failed");
-    assertEquals(result.s, emptyStruct.s, "Empty CustomStruct.s serialization failed");
-    
-    // Test CustomStruct with values
-    const filledStruct = new CustomStruct([0.1, 0.2, -1000.55], 32, "test");
-    bytes = zserialize(filledStruct);
-    result = zdeserialize(ZD.object(CustomStruct), bytes);
-    assertEquals(result.vd, filledStruct.vd, "CustomStruct.vd serialization failed");
-    assertEquals(result.i, filledStruct.i, "CustomStruct.i serialization failed");
-    assertEquals(result.s, filledStruct.s, "CustomStruct.s serialization failed");
-});
-
-/**
- * 3. TYPED ARRAYS TESTS
- */
-Deno.test("Serialization - Int Typed Arrays", () => {
-    // Test Int8Array
-    const int8Data = new Int8Array([-128, -100, -50, 0, 50, 100, 127]);
-    let bytes = zserialize(int8Data, ZS.int8array());
-    const result = zdeserialize(ZD.int8array(), bytes);
-    assert(result instanceof Int8Array, "Result should be Int8Array");
-    assertEquals(Array.from(result), Array.from(int8Data), "Int8Array serialization failed");
-
-    // Test Int16Array
-    const int16Data = new Int16Array([100, -200, 300, -30000]);
-    bytes = zserialize(int16Data, ZS.int16array());
-    const int16Result = zdeserialize(ZD.int16array(), bytes);
-    assert(int16Result instanceof Int16Array, "Result should be Int16Array");
-    assertEquals(Array.from(int16Result), Array.from(int16Data), "Int16Array serialization failed");
-    
-    // Test Int32Array
-    const int32Data = new Int32Array([100000, -200000, 300000, -2000000000]);
-    bytes = zserialize(int32Data, ZS.int32array());
-    const int32Result = zdeserialize(ZD.int32array(), bytes);
-    assert(int32Result instanceof Int32Array, "Result should be Int32Array");
-    assertEquals(Array.from(int32Result), Array.from(int32Data), "Int32Array serialization failed");
-    
-    // Test BigInt64Array
-    const bigInt64Data = new BigInt64Array([100000n, -200000n, 300000n, -2000000000n]);
-    bytes = zserialize(bigInt64Data, ZS.bigint64array());
-    const bigResult = zdeserialize(ZD.bigint64array(), bytes);
-    assert(bigResult instanceof BigInt64Array, "Result should be BigInt64Array");
-    assertEquals(Array.from(bigResult), Array.from(bigInt64Data), "BigInt64Array serialization failed");
-});
-
-Deno.test("Serialization - Uint Typed Arrays", () => {
-    // Test Uint8Array
-    const uint8Data = new Uint8Array([0, 127, 255]);
-    let bytes = zserialize(uint8Data, ZS.uint8array());
-    const result = zdeserialize(ZD.uint8array(), bytes);
-    assert(result instanceof Uint8Array, "Result should be Uint8Array");
-    assertEquals(Array.from(result), Array.from(uint8Data), "Uint8Array serialization failed");
-
-    // Test Uint16Array
-    const uint16Data = new Uint16Array([100, 200, 300, 40000]);
-    bytes = zserialize(uint16Data, ZS.uint16array());
-    const uint16Result = zdeserialize(ZD.uint16array(), bytes);
-    assert(uint16Result instanceof Uint16Array, "Result should be Uint16Array");
-    assertEquals(Array.from(uint16Result), Array.from(uint16Data), "Uint16Array serialization failed");
-
-    // Test Uint32Array
-    const uint32Data = new Uint32Array([100000, 200000, 300000, 4000000000]);
-    bytes = zserialize(uint32Data, ZS.uint32array());
-    const uint32Result = zdeserialize(ZD.uint32array(), bytes);
-    assert(uint32Result instanceof Uint32Array, "Result should be Uint32Array");
-    assertEquals(Array.from(uint32Result), Array.from(uint32Data), "Uint32Array serialization failed");
-    
-    // Test BigUint64Array
-    const bigUint64Data = new BigUint64Array([100000n, 200000n, 300000n, 4000000000n]);
-    bytes = zserialize(bigUint64Data, ZS.biguint64array());
-    const bigResult = zdeserialize(ZD.biguint64array(), bytes);
-    assert(bigResult instanceof BigUint64Array, "Result should be BigUint64Array");
-    assertEquals(Array.from(bigResult), Array.from(bigUint64Data), "BigUint64Array serialization failed");
-});
-
-Deno.test("Serialization - Float Typed Arrays", () => {
-    // Test Float32Array
-    const float32Data = new Float32Array([1.5, -2.25, 3.75, -4.125]);
-    const float32Bytes = zserialize(float32Data, ZS.float32array());
-    const float32Result = zdeserialize(ZD.float32array(), float32Bytes);
-    assert(float32Result instanceof Float32Array, "Result should be Float32Array");
-    assert(compareFloatArrays(float32Result, float32Data), "Float32Array serialization failed");
-    
-    // Test Float64Array
-    const float64Data = new Float64Array([1.12345, -2.34567, 3.45678, -4.56789]);
-    const float64Bytes = zserialize(float64Data, ZS.float64array());
-    const float64Result = zdeserialize(ZD.float64array(), float64Bytes);
-    assert(float64Result instanceof Float64Array, "Result should be Float64Array");
-    assert(compareFloatArrays(float64Result, float64Data), "Float64Array serialization failed");
-});
-
-Deno.test("Serialization - TypedArray Auto Detection", () => {
-    // Test auto detection for Uint8Array
-    const uint8Data = new Uint8Array([0, 127, 255]);
-    const uint8Bytes = zserialize(uint8Data); // No explicit serializer tag
-    const uint8Result = zdeserialize(ZD.uint8array(), uint8Bytes);
-    assert(uint8Result instanceof Uint8Array, "Result should be Uint8Array");
-    assertEquals(Array.from(uint8Result), Array.from(uint8Data), "Auto-detected Uint8Array serialization failed");
-
-    // Test auto detection for Int8Array
-    const int8Data = new Int8Array([-128, -100, -50, 0, 50, 100, 127]);
-    const int8Bytes = zserialize(int8Data); // No explicit serializer tag
-    const int8Result = zdeserialize(ZD.int8array(), int8Bytes);
-    assert(int8Result instanceof Int8Array, "Result should be Int8Array");
-    assertEquals(Array.from(int8Result), Array.from(int8Data), "Auto-detected Int8Array serialization failed");
-
-    // Test auto detection for Float32Array
-    const float32Data = new Float32Array([1.5, -2.25, 3.75, -4.125]);
-    const float32Bytes = zserialize(float32Data); // No explicit serializer tag
-    const float32Result = zdeserialize(ZD.float32array(), float32Bytes);
-    assert(float32Result instanceof Float32Array, "Result should be Float32Array");
-    assert(compareFloatArrays(float32Result, float32Data), "Auto-detected Float32Array serialization failed");
-});
-
-Deno.test("Serialization - TypedArray Collections", () => {
-    // Test array of TypedArrays
-    const typedArrayCollection = [
-        new Uint16Array([1, 2, 3]),
-        new Uint16Array([4, 5, 6]),
-        new Uint16Array([7, 8, 9])
-    ];
-    
-    const collectionBytes = zserialize(typedArrayCollection, ZS.array(ZS.uint16array()));
-    const collectionResult = zdeserialize(ZD.array(ZD.uint16array()), collectionBytes);
-    
-    assert(Array.isArray(collectionResult), "Result should be an Array");
-    assertEquals(collectionResult.length, typedArrayCollection.length, "Collection length mismatch");
-    
-    for (let i = 0; i < collectionResult.length; i++) {
-        assert(collectionResult[i] instanceof Uint16Array, `Item ${i} should be Uint16Array`);
-        assertEquals(
-            Array.from(collectionResult[i]), 
-            Array.from(typedArrayCollection[i]),
-            `Item ${i} data mismatch`
-        );
-    }
-    
-    // Test Map with TypedArray keys and values
-    const typedArrayMap = new Map<Uint8Array, Float32Array>();
-    typedArrayMap.set(
-        new Uint8Array([1, 2, 3]), 
-        new Float32Array([1.1, 2.2, 3.3])
-    );
-    typedArrayMap.set(
-        new Uint8Array([4, 5, 6]), 
-        new Float32Array([4.4, 5.5, 6.6])
-    );
-    
-    const mapBytes = zserialize(typedArrayMap, ZS.map(ZS.uint8array(), ZS.float32array()));
-    const mapResult = zdeserialize(ZD.map(ZD.uint8array(), ZD.float32array()), mapBytes);
-    
-    assert(mapResult instanceof Map, "Result should be a Map");
-    assertEquals(mapResult.size, typedArrayMap.size, "Map size mismatch");
-    
-    // Compare map entries - need to convert to strings for comparison since TypedArrays can't be directly compared
-    for (const [key, value] of typedArrayMap.entries()) {
-        // Find matching key in result
-        let foundMatch = false;
-        for (const [resultKey, resultValue] of mapResult.entries()) {
-            if (Array.from(resultKey).toString() === Array.from(key).toString()) {
-                foundMatch = true;
-                assert(compareFloatArrays(resultValue, value), "Map value mismatch");
-                break;
-            }
-        }
-        assert(foundMatch, `No matching key found for ${Array.from(key)}`);
-    }
-    
-    // Test mixed collection with different TypedArrays
-    const mixedCollection = {
-        uint16: new Uint16Array([1, 2]),
-        int32: new Int32Array([3, 4]),
-        float64: new Float64Array([5.5, 6.6])
-    };
-    
-    const mixed = new MixedTypedArrayCollection(
-        mixedCollection.uint16,
-        mixedCollection.int32,
-        mixedCollection.float64
-    );
-    
-    const mixedBytes = zserialize(mixed);
-    const mixedResult = zdeserialize(ZD.object(MixedTypedArrayCollection), mixedBytes);
-    
-    assert(mixedResult instanceof MixedTypedArrayCollection, "Result should be a MixedTypedArrayCollection");
-    assertEquals(
-        Array.from(mixedResult.uint16), 
-        Array.from(mixedCollection.uint16),
-        "uint16 mismatch"
-    );
-    assertEquals(
-        Array.from(mixedResult.int32), 
-        Array.from(mixedCollection.int32),
-        "int32 mismatch"
-    );
-    assert(
-        compareFloatArrays(mixedResult.float64, mixedCollection.float64),
-        "float64 mismatch"
-    );
-});
-
-/**
- * 4. LOW-LEVEL API TESTS
- */
-Deno.test("Serialization - Direct Serializer/Deserializer API", () => {
-    // Test direct API for numbers
-    const numbers = [-128, 0, 127, 255];
-    for (const val of numbers) {
-        const serializer = new ZBytesSerializer();
-        
-        if (val >= 0 && val <= 255) {
-            serializer.serialize(val, ZS.number(NumberFormat.Uint8));
-        } else {
-            serializer.serialize(val, ZS.number(NumberFormat.Int8));
-        }
-        
-        const bytes = serializer.finish();
-        const deserializer = new ZBytesDeserializer(bytes);
-        
-        let result;
-        if (val >= 0 && val <= 255) {
-            result = deserializer.deserialize(ZD.number(NumberFormat.Uint8));
-        } else {
-            result = deserializer.deserialize(ZD.number(NumberFormat.Int8));
-        }
-        
-        assertEquals(result, val, `Direct serialization API failed for value: ${val}`);
-    }
-    
-    // Test direct API for typed arrays
-    const int8Data = new Int8Array([-128, -64, 0, 64, 127]);
-    const serializer = new ZBytesSerializer();
-    serializer.serialize(int8Data, ZS.int8array());
-    const bytes = serializer.finish();
-    
-    const deserializer = new ZBytesDeserializer(bytes);
-    const result = deserializer.deserialize(ZD.int8array());
-    
-    assert(result instanceof Int8Array, "Result should be Int8Array");
-    assertEquals(Array.from(result), Array.from(int8Data), "Direct API for Int8Array failed");
-    
-    // Additional test for specific edge cases of Int8 and Uint8
-    for (const val of [-128, -1, 0, 1, 127]) {
-        const s1 = new ZBytesSerializer();
-        s1.serialize(val, ZS.number(NumberFormat.Int8));
-        const b1 = s1.finish();
-        
-        const d1 = new ZBytesDeserializer(b1);
-        const r1 = d1.deserialize(ZD.number(NumberFormat.Int8));
-        assertEquals(r1, val, `Int8 direct serialization failed for ${val}`);
-    }
-    
-    for (const val of [0, 1, 127, 128, 254, 255]) {
-        const s2 = new ZBytesSerializer();
-        s2.serialize(val, ZS.number(NumberFormat.Uint8));
-        const b2 = s2.finish();
-        
-        const d2 = new ZBytesDeserializer(b2);
-        const r2 = d2.deserialize(ZD.number(NumberFormat.Uint8));
-        assertEquals(r2, val, `Uint8 direct serialization failed for ${val}`);
-    }
-});
-
-// Test targeting specific Uint8 serialization/deserialization mechanisms
-Deno.test("Serialization - Specialized Uint8 Functions", () => {
-    // Test all edge cases for Uint8 values
-    const uint8Values = [0, 1, 127, 128, 254, 255];
-    
-    // Direct use of serializer/deserializer for Uint8
-    for (const val of uint8Values) {
-        const serializer = new ZBytesSerializer();
-        serializer.serialize(val, ZS.number(NumberFormat.Uint8));
-        const bytes = serializer.finish();
-        
-        const deserializer = new ZBytesDeserializer(bytes);
-        const result = deserializer.deserialize(ZD.number(NumberFormat.Uint8));
-        
-        assertEquals(result, val, `Specialized Uint8 serialization failed for value: ${val}`);
-    }
-    
-    // Test direct serialization of Uint8Array with specific values
-    const uint8Data = new Uint8Array([0, 128, 255]);
-    const bytes = zserialize(uint8Data, ZS.uint8array());
-    const result = zdeserialize(ZD.uint8array(), bytes);
-    
-    assert(result instanceof Uint8Array, "Result should be Uint8Array");
-    assertEquals(Array.from(result), Array.from(uint8Data), "Uint8Array specialized serialization failed");
-});
-
-// Extended test focusing on Int8Array edge cases and collections
-Deno.test("Serialization - Comprehensive Int8Array Coverage", () => {
-    // Create an Int8Array with the full range of possible values
-    const _int8Data = new Int8Array([-128, -100, -50, 0, 50, 100, 127]);
-    
-    // Test array of Int8Arrays to ensure collection handling
-    const int8ArrayCollection = [
-        new Int8Array([-10, 20, -30]),
-        new Int8Array([40, -50, 60])
-    ];
-    
-    const collectionBytes = zserialize(int8ArrayCollection, ZS.array(ZS.int8array()));
-    const collectionResult = zdeserialize(ZD.array(ZD.int8array()), collectionBytes);
-    
-    assert(Array.isArray(collectionResult), "Result should be an Array");
-    assertEquals(collectionResult.length, int8ArrayCollection.length, "Collection length mismatch");
-    
-    for (let i = 0; i < collectionResult.length; i++) {
-        assert(collectionResult[i] instanceof Int8Array, `Item ${i} should be Int8Array`);
-        assertEquals(
-            Array.from(collectionResult[i]),
-            Array.from(int8ArrayCollection[i]),
-            `Item ${i} data mismatch`
-        );
-    }
-});
-
-/**
- * 5. EDGE CASES AND ERROR HANDLING
- */
-Deno.test("Serialization - Edge Cases and Error Handling", () => {
-    // Test serializing a value larger than MAX_SAFE_INTEGER
-    const tooLargeNum = 9876543210123456789n; // Larger than 53 bits
-    const bytes = zserialize(tooLargeNum, ZS.bigint(BigIntFormat.Uint64));
-    
-    // Should deserialize as BigInt without error
-    const bigintResult = zdeserialize(ZD.bigint(BigIntFormat.Uint64), bytes);
-    assertEquals(bigintResult, tooLargeNum, "BigInt serialization failed");
-    
-    // Should throw when trying to deserialize as Number
-    let threw = false;
-    try {
-        zdeserialize(ZD.number(NumberFormat.Uint64), bytes);
-    } catch (e) {
-        threw = true;
-        assert(e instanceof Error && e.message.includes("exceeds the safe range"), 
-               "Error message for large number deserialization is incorrect");
-    }
-    assert(threw, "Should throw when deserializing bigint > MAX_SAFE_INTEGER as number");
-    
-    // Test edge cases for Uint64/Int64 within safe integer range
-    // This specifically targets Uint64 serialization functions
-    const uint64MaxSafe = Number.MAX_SAFE_INTEGER;
-    const uint64Bytes = zserialize(uint64MaxSafe, ZS.number(NumberFormat.Uint64));
-    const uint64Result = zdeserialize(ZD.number(NumberFormat.Uint64), uint64Bytes);
-    assertEquals(uint64Result, uint64MaxSafe, "Uint64 max safe integer serialization failed");
-    
-    // Test with a range of values for better coverage of Uint64
-    const uint64Values = [0, 100, 1000000, 9007199254740991]; // MAX_SAFE_INTEGER
-    for (const val of uint64Values) {
-        const bytes = zserialize(val, ZS.number(NumberFormat.Uint64));
-        const result = zdeserialize(ZD.number(NumberFormat.Uint64), bytes);
-        assertEquals(result, val, `Uint64 serialization failed for value: ${val}`);
-    }
-    
-    // Test Int64 minimum safe integer
-    const int64MinSafe = -Number.MAX_SAFE_INTEGER;
-    const int64Bytes = zserialize(int64MinSafe, ZS.number(NumberFormat.Int64));
-    const int64Result = zdeserialize(ZD.number(NumberFormat.Int64), int64Bytes);
-    assertEquals(int64Result, int64MinSafe, "Int64 min safe integer serialization failed");
-});
-
-// Test focusing on handling specific serialization edge cases 
-Deno.test("Serialization - Additional Edge Cases", () => {
-    // Test for edge case handling in Uint8 serialization
-    const edgeUint8Values = [0, 1, 128, 255];
-    for (const val of edgeUint8Values) {
-        const bytes = zserialize(val, ZS.number(NumberFormat.Uint8));
-        const result = zdeserialize(ZD.number(NumberFormat.Uint8), bytes);
-        assertEquals(result, val, `Edge case Uint8 serialization failed for value: ${val}`);
-    }
-    
-    // Test for edge case handling in Int8 serialization
-    const edgeInt8Values = [-128, -1, 0, 1, 127];
-    for (const val of edgeInt8Values) {
-        const bytes = zserialize(val, ZS.number(NumberFormat.Int8));
-        const result = zdeserialize(ZD.number(NumberFormat.Int8), bytes);
-        assertEquals(result, val, `Edge case Int8 serialization failed for value: ${val}`);
-    }
-    
-    // Test specific Int32 values
-    const int32Tests = [1234566, -49245];
-    for (const val of int32Tests) {
-        const bytes = zserialize(val, ZS.number(NumberFormat.Int32));
-        const result = zdeserialize(ZD.number(NumberFormat.Int32), bytes);
-        assertEquals(result, val, "Int32 specific value serialization failed");
-    }
-    
-    // Test number/string combination serializations
-    const num1 = 500;
-    const num2 = 1234.0;
-    const testStr = "test";
-    
-    const bytes1 = zserialize(num1, ZS.number());
-    const bytes2 = zserialize(num2, ZS.number());
-    const bytes3 = zserialize(testStr, ZS.string());
-    
-    assertEquals(zdeserialize(ZD.number(), bytes1), num1, "Simple number serialization failed");
-    assertEquals(zdeserialize(ZD.number(), bytes2), num2, "Float number serialization failed");
-    assertEquals(zdeserialize(ZD.string(), bytes3), testStr, "String serialization with numbers failed");
-    
-    // Test zero-length arrays
-    const emptyUint8Array = new Uint8Array(0);
-    const emptyUint8Bytes = zserialize(emptyUint8Array, ZS.uint8array());
-    const emptyUint8Result = zdeserialize(ZD.uint8array(), emptyUint8Bytes);
-    assert(emptyUint8Result instanceof Uint8Array, "Result should be Uint8Array");
-    assertEquals(emptyUint8Result.length, 0, "Empty Uint8Array serialization failed");
-    
-    // Test empty Int8Array
-    const emptyInt8Array = new Int8Array(0);
-    const emptyInt8Bytes = zserialize(emptyInt8Array, ZS.int8array());
-    const emptyInt8Result = zdeserialize(ZD.int8array(), emptyInt8Bytes);
-    assert(emptyInt8Result instanceof Int8Array, "Result should be Int8Array");
-    assertEquals(emptyInt8Result.length, 0, "Empty Int8Array serialization failed");
-});
-
-/**
- * 6. BIG-ENDIAN AND SYSTEM-SPECIFIC TESTS
- */
-Deno.test("Serialization - Big Endian Simulation", () => {
-    // Test TypedArray serialization on non-little-endian systems
-    // This tests the else branches in serialization methods
-    
-    // Create arrays that would trigger different serialization paths
-    const uint16Data = new Uint16Array([0x1234, 0x5678, 0xABCD]);
-    const uint32Data = new Uint32Array([0x12345678, 0xABCDEF01]);
-    const bigUint64Data = new BigUint64Array([0x123456789ABCDEFn, 0xFEDCBA9876543210n]);
-    
-    const int16Data = new Int16Array([-1234, 5678, -9876]);
-    const int32Data = new Int32Array([-123456789, 987654321]);
-    const bigInt64Data = new BigInt64Array([-0x123456789ABCDEFn, 0x7EDCBa9876543210n]);
-    
-    const float32Data = new Float32Array([3.14159, -2.71828, 1.41421]);
-    const float64Data = new Float64Array([Math.PI, Math.E, Math.SQRT2]);
-    
-    // Force serialization through non-optimized paths by manipulating endianness detection
-    // We'll serialize and deserialize to ensure both paths work
-    
-    // Test Uint16Array
-    let bytes = zserialize(uint16Data, ZS.uint16array());
-    const uint16Result = zdeserialize(ZD.uint16array(), bytes);
-    assertEquals(Array.from(uint16Result), Array.from(uint16Data), "Uint16Array big-endian serialization failed");
-    
-    // Test Uint32Array
-    bytes = zserialize(uint32Data, ZS.uint32array());
-    const uint32Result = zdeserialize(ZD.uint32array(), bytes);
-    assertEquals(Array.from(uint32Result), Array.from(uint32Data), "Uint32Array big-endian serialization failed");
-    
-    // Test BigUint64Array
-    bytes = zserialize(bigUint64Data, ZS.biguint64array());
-    const bigUint64Result = zdeserialize(ZD.biguint64array(), bytes);
-    assertEquals(Array.from(bigUint64Result), Array.from(bigUint64Data), "BigUint64Array big-endian serialization failed");
-    
-    // Test Int16Array
-    bytes = zserialize(int16Data, ZS.int16array());
-    const int16Result = zdeserialize(ZD.int16array(), bytes);
-    assertEquals(Array.from(int16Result), Array.from(int16Data), "Int16Array big-endian serialization failed");
-    
-    // Test Int32Array
-    bytes = zserialize(int32Data, ZS.int32array());
-    const int32Result = zdeserialize(ZD.int32array(), bytes);
-    assertEquals(Array.from(int32Result), Array.from(int32Data), "Int32Array big-endian serialization failed");
-    
-    // Test BigInt64Array
-    bytes = zserialize(bigInt64Data, ZS.bigint64array());
-    const bigInt64Result = zdeserialize(ZD.bigint64array(), bytes);
-    assertEquals(Array.from(bigInt64Result), Array.from(bigInt64Data), "BigInt64Array big-endian serialization failed");
-    
-    // Test Float32Array
-    bytes = zserialize(float32Data, ZS.float32array());
-    const float32Result = zdeserialize(ZD.float32array(), bytes);
-    assert(compareFloatArrays(float32Result, float32Data), "Float32Array big-endian serialization failed");
-    
-    // Test Float64Array
-    bytes = zserialize(float64Data, ZS.float64array());
-    const float64Result = zdeserialize(ZD.float64array(), bytes);
-    assert(compareFloatArrays(float64Result, float64Data), "Float64Array big-endian serialization failed");
-});
-
-Deno.test("Serialization - Error Handling and Bounds", () => {
-    // Test various error conditions to improve branch coverage
-    
-    // Test array index out of bounds during deserialization
-    const serializer = new ZBytesSerializer();
-    serializer.serializeString("test");
-    const bytes = serializer.finish();
-    
-    const deserializer = new ZBytesDeserializer(bytes);
-    deserializer.deserializeString(); // This should work
-    
-    // Try to read more data than available
-    let errorThrown = false;
-    try {
-        deserializer.deserializeString(); // This should fail
-    } catch (e) {
-        errorThrown = true;
-        assert((e as Error).message.includes("out of bounds"), "Expected out of bounds error");
-    }
-    assert(errorThrown, "Expected error for reading beyond buffer bounds");
-    
-    // Test sequence length overflow
-    const largeSerializer = new ZBytesSerializer();
-    // Test with a very large sequence that could cause overflow
-    largeSerializer.writeSequenceLength(Number.MAX_SAFE_INTEGER + 1);
-    
-    // Test invalid boolean values - create manually by serializing an invalid value
-    const boolSerializer = new ZBytesSerializer();
-    boolSerializer.serializeNumberUint8(5); // Invalid boolean value (not 0 or 1)
-    const invalidBoolBytes = boolSerializer.finish();
-    
-    errorThrown = false;
-    try {
-        zdeserialize(ZD.boolean(), invalidBoolBytes);
-    } catch (e) {
-        errorThrown = true;
-        assert((e as Error).message.includes("Unexpected boolean value"), "Expected boolean error");
-    }
-    assert(errorThrown, "Expected error for invalid boolean value");
-    
-    // Test payload with extra bytes - create a valid string with extra data
-    const extraBytesSerializer = new ZBytesSerializer();
-    extraBytesSerializer.serializeString("test");
-    extraBytesSerializer.serializeNumberUint8(1); // Extra byte
-    extraBytesSerializer.serializeNumberUint8(2); // More extra bytes
-    extraBytesSerializer.serializeNumberUint8(3);
-    const extraBytes = extraBytesSerializer.finish();
-    
-    errorThrown = false;
-    try {
-        zdeserialize(ZD.string(), extraBytes);
-    } catch (e) {
-        errorThrown = true;
-        assert((e as Error).message.includes("more bytes than required"), "Expected extra bytes error");
-    }
-    assert(errorThrown, "Expected error for extra bytes in payload");
-});
-
-Deno.test("Serialization - NumberUint8 Specific Edge Cases", () => {
-    // Test the specific serializeNumberUint8 method that shows in coverage
-    // Note: deserializeNumberUint8 actually returns signed int8, so we test accordingly
-    
-    const testValues = [
-        { input: 0, expected: 0 },
-        { input: 127, expected: 127 },
-        { input: 128, expected: -128 }, // Uint8 128 becomes Int8 -128
-        { input: 255, expected: -1 }   // Uint8 255 becomes Int8 -1
-    ];
-    
-    for (const testCase of testValues) {
-        const serializer = new ZBytesSerializer();
-        serializer.serializeNumberUint8(testCase.input);
-        const bytes = serializer.finish();
-        
-        const deserializer = new ZBytesDeserializer(bytes);
-        const result = deserializer.deserializeNumberUint8();
-        assertEquals(result, testCase.expected, 
-                    `NumberUint8 serialization failed for ${testCase.input} -> ${testCase.expected}`);
-    }
-});
-
-Deno.test("Serialization - Default Tag Detection Edge Cases", () => {
-    // Test getDefaultSerializationTag branches that aren't covered
-    
-    // Test boolean
-    const boolData = true;
-    const boolBytes = zserialize(boolData);
-    const boolResult = zdeserialize(ZD.boolean(), boolBytes);
-    assertEquals(boolResult, boolData, `Auto serialization failed for boolean`);
-    
-    // Test bigint
-    const bigintData = 42n;
-    const bigintBytes = zserialize(bigintData);
-    const bigintResult = zdeserialize(ZD.bigint(), bigintBytes);
-    assertEquals(bigintResult, bigintData, `Auto serialization failed for bigint`);
-    
-    // Test various typed arrays
-    const uint16Data = new Uint16Array([1, 2]);
-    const uint16Bytes = zserialize(uint16Data);
-    const uint16Result = zdeserialize(ZD.uint16array(), uint16Bytes);
-    assertEquals(Array.from(uint16Result), Array.from(uint16Data), 
-                `Auto serialization failed for Uint16Array`);
-    
-    const uint32Data = new Uint32Array([1, 2]);
-    const uint32Bytes = zserialize(uint32Data);
-    const uint32Result = zdeserialize(ZD.uint32array(), uint32Bytes);
-    assertEquals(Array.from(uint32Result), Array.from(uint32Data), 
-                `Auto serialization failed for Uint32Array`);
-    
-    const bigUint64Data = new BigUint64Array([1n, 2n]);
-    const bigUint64Bytes = zserialize(bigUint64Data);
-    const bigUint64Result = zdeserialize(ZD.biguint64array(), bigUint64Bytes);
-    assertEquals(Array.from(bigUint64Result), Array.from(bigUint64Data), 
-                `Auto serialization failed for BigUint64Array`);
-    
-    const int16Data = new Int16Array([1, 2]);
-    const int16Bytes = zserialize(int16Data);
-    const int16Result = zdeserialize(ZD.int16array(), int16Bytes);
-    assertEquals(Array.from(int16Result), Array.from(int16Data), 
-                `Auto serialization failed for Int16Array`);
-    
-    const int32Data = new Int32Array([1, 2]);
-    const int32Bytes = zserialize(int32Data);
-    const int32Result = zdeserialize(ZD.int32array(), int32Bytes);
-    assertEquals(Array.from(int32Result), Array.from(int32Data), 
-                `Auto serialization failed for Int32Array`);
-    
-    const bigInt64Data = new BigInt64Array([1n, 2n]);
-    const bigInt64Bytes = zserialize(bigInt64Data);
-    const bigInt64Result = zdeserialize(ZD.bigint64array(), bigInt64Bytes);
-    assertEquals(Array.from(bigInt64Result), Array.from(bigInt64Data), 
-                `Auto serialization failed for BigInt64Array`);
-    
-    const float64Data = new Float64Array([1.0, 2.0]);
-    const float64Bytes = zserialize(float64Data);
-    const float64Result = zdeserialize(ZD.float64array(), float64Bytes);
-    assertEquals(Array.from(float64Result), Array.from(float64Data), 
-                `Auto serialization failed for Float64Array`);
-    
-    // Test array with undefined tag detection
-    const mixedArray = [1, 2, 3];
-    const arrayBytes = zserialize(mixedArray);
-    const arrayResult = zdeserialize(ZD.array(ZD.number()), arrayBytes);
-    assertEquals(arrayResult, mixedArray, "Array auto-tag detection failed");
-    
-    // Test map with undefined key/value tags
-    const testMap = new Map([["key1", 42], ["key2", 84]]);
-    const mapBytes = zserialize(testMap);
-    const mapResult = zdeserialize(ZD.map(ZD.string(), ZD.number()), mapBytes);
-    assertEquals(Array.from(mapResult.entries()), Array.from(testMap.entries()), "Map auto-tag detection failed");
-});
-
-Deno.test("Serialization - Non-Serializable Type Error", () => {
-    // Test the error path for non-ZSerializeable types
-    const invalidData = Symbol("test"); // Symbols are not serializable
-    
-    let errorThrown = false;
-    try {
-        // This should trigger the "Non-ZSerializeable type" error
-        // @ts-ignore - Intentionally testing invalid type
-        zserialize(invalidData);
-    } catch (e) {
-        errorThrown = true;
-        assert((e as Error).message.includes("Non-ZSerializeable type") || 
-               (e as Error).message.includes("Cannot serialize"), 
-               "Expected non-serializable type error");
-    }
-    assert(errorThrown, "Expected error for non-serializable type");
-    
-    // Test with a plain object that doesn't implement ZSerializeable
-    const plainObject = { x: 1, y: 2 };
-    errorThrown = false;
-    try {
-        // @ts-ignore - Intentionally testing invalid type
-        zserialize(plainObject);
-    } catch (e) {
-        errorThrown = true;
-        assert((e as Error).message.includes("Non-ZSerializeable type") || 
-               (e as Error).message.includes("Cannot serialize"), 
-               "Expected non-serializable object error");
-    }
-    assert(errorThrown, "Expected error for plain object without ZSerializeable interface");
-});
-
-Deno.test("Serialization - Edge Cases for getDefaultSerializationTag Branches", () => {
-    // Test empty array (should have undefined tag)
-    const emptyArray: number[] = [];
-    const emptyArrayBytes = zserialize(emptyArray);
-    const emptyArrayResult = zdeserialize(ZD.array(ZD.number()), emptyArrayBytes);
-    assertEquals(emptyArrayResult, emptyArray, "Empty array serialization failed");
-    
-    // Test empty map (should have undefined key/value tags)
-    const emptyMap = new Map<string, number>();
-    const emptyMapBytes = zserialize(emptyMap);
-    const emptyMapResult = zdeserialize(ZD.map(ZD.string(), ZD.number()), emptyMapBytes);
-    assertEquals(Array.from(emptyMapResult.entries()), Array.from(emptyMap.entries()), "Empty map serialization failed");
-    
-    // Test map where entries().next() returns valid data but edge case handling
-    const singleEntryMap = new Map([["test", 123]]);
-    const singleMapBytes = zserialize(singleEntryMap);
-    const singleMapResult = zdeserialize(ZD.map(ZD.string(), ZD.number()), singleMapBytes);
-    assertEquals(Array.from(singleMapResult.entries()), Array.from(singleEntryMap.entries()), "Single entry map failed");
-    
-    // Test different data types to ensure all branches are hit
-    const boolTest = true;
-    const boolBytes = zserialize(boolTest);
-    const boolResult = zdeserialize(ZD.boolean(), boolBytes);
-    assertEquals(boolResult, boolTest, "boolean serialization failed");
-    
-    const bigintTest = 123n;
-    const bigintBytes = zserialize(bigintTest);
-    const bigintResult = zdeserialize(ZD.bigint(), bigintBytes);
-    assertEquals(bigintResult, bigintTest, "bigint serialization failed");
-    
-    const stringTest = "test";
-    const stringBytes = zserialize(stringTest);
-    const stringResult = zdeserialize(ZD.string(), stringBytes);
-    assertEquals(stringResult, stringTest, "string serialization failed");
-    
-    const numberTest = 42.5;
-    const numberBytes = zserialize(numberTest);
-    const numberResult = zdeserialize(ZD.number(), numberBytes);
-    assertEquals(numberResult, numberTest, "number serialization failed");
-    
-    const uint8Test = new Uint8Array([1, 2, 3]);
-    const uint8Bytes = zserialize(uint8Test);
-    const uint8Result = zdeserialize(ZD.uint8array(), uint8Bytes);
-    assertEquals(Array.from(uint8Result), Array.from(uint8Test), "Uint8Array serialization failed");
-    
-    const float32Test = new Float32Array([1.0, 2.0]);
-    const float32Bytes = zserialize(float32Test);
-    const float32Result = zdeserialize(ZD.float32array(), float32Bytes);
-    assertEquals(Array.from(float32Result), Array.from(float32Test), "Float32Array serialization failed");
-});
-
-// Test targeting very specific uncovered branches to improve coverage
-Deno.test("Serialization - Complete Branch Coverage for getDefaultSerializationTag", () => {
-    // Test custom object implementing ZSerializeable (first branch in getDefaultSerializationTag)
-    class TestSerializeable implements ZSerializeable, ZDeserializeable {
-        public value: number = 42;
-        
-        serializeWithZSerializer(serializer: ZBytesSerializer): void {
-            serializer.serialize(this.value, ZS.number());
-        }
-        
-        deserializeWithZDeserializer(deserializer: ZBytesDeserializer): void {
-            this.value = deserializer.deserialize(ZD.number());
-        }
-    }
-    
-    const testObj = new TestSerializeable();
-    const objBytes = zserialize(testObj);
-    const objResult = zdeserialize(ZD.object(TestSerializeable), objBytes);
-    assertEquals(objResult.value, testObj.value, "ZSerializeable object serialization failed");
-    
-    // Test all basic types to ensure complete branch coverage
-    const testNumber = 123.456;
-    const numBytes = zserialize(testNumber);
-    const numResult = zdeserialize(ZD.number(), numBytes);
-    assertEquals(numResult, testNumber, "Number auto-tag failed");
-    
-    const testBigint = 9007199254740991n;
-    const bigintBytes = zserialize(testBigint);
-    const bigintResult = zdeserialize(ZD.bigint(), bigintBytes);
-    assertEquals(bigintResult, testBigint, "Bigint auto-tag failed");
-    
-    const testString = "hello world";
-    const stringBytes = zserialize(testString);
-    const stringResult = zdeserialize(ZD.string(), stringBytes);
-    assertEquals(stringResult, testString, "String auto-tag failed");
-    
-    const testBoolean = false;
-    const boolBytes = zserialize(testBoolean);
-    const boolResult = zdeserialize(ZD.boolean(), boolBytes);
-    assertEquals(boolResult, testBoolean, "Boolean auto-tag failed");
-    
-    // Test all TypedArray types for complete branch coverage
-    const uint8Test = new Uint8Array([1, 2, 3]);
-    const uint8Bytes = zserialize(uint8Test);
-    const uint8Result = zdeserialize(ZD.uint8array(), uint8Bytes);
-    assertEquals(Array.from(uint8Result), Array.from(uint8Test), "Uint8Array auto-tag failed");
-    
-    const uint16Test = new Uint16Array([256, 512]);
-    const uint16Bytes = zserialize(uint16Test);
-    const uint16Result = zdeserialize(ZD.uint16array(), uint16Bytes);
-    assertEquals(Array.from(uint16Result), Array.from(uint16Test), "Uint16Array auto-tag failed");
-    
-    const uint32Test = new Uint32Array([65536, 131072]);
-    const uint32Bytes = zserialize(uint32Test);
-    const uint32Result = zdeserialize(ZD.uint32array(), uint32Bytes);
-    assertEquals(Array.from(uint32Result), Array.from(uint32Test), "Uint32Array auto-tag failed");
-    
-    const bigUint64Test = new BigUint64Array([1n, 2n, 3n]);
-    const bigUint64Bytes = zserialize(bigUint64Test);
-    const bigUint64Result = zdeserialize(ZD.biguint64array(), bigUint64Bytes);
-    assertEquals(Array.from(bigUint64Result), Array.from(bigUint64Test), "BigUint64Array auto-tag failed");
-    
-    const int8Test = new Int8Array([-1, 0, 1]);
-    const int8Bytes = zserialize(int8Test);
-    const int8Result = zdeserialize(ZD.int8array(), int8Bytes);
-    assertEquals(Array.from(int8Result), Array.from(int8Test), "Int8Array auto-tag failed");
-    
-    const int16Test = new Int16Array([-256, 0, 256]);
-    const int16Bytes = zserialize(int16Test);
-    const int16Result = zdeserialize(ZD.int16array(), int16Bytes);
-    assertEquals(Array.from(int16Result), Array.from(int16Test), "Int16Array auto-tag failed");
-    
-    const int32Test = new Int32Array([-65536, 0, 65536]);
-    const int32Bytes = zserialize(int32Test);
-    const int32Result = zdeserialize(ZD.int32array(), int32Bytes);
-    assertEquals(Array.from(int32Result), Array.from(int32Test), "Int32Array auto-tag failed");
-    
-    const bigInt64Test = new BigInt64Array([-1n, 0n, 1n]);
-    const bigInt64Bytes = zserialize(bigInt64Test);
-    const bigInt64Result = zdeserialize(ZD.bigint64array(), bigInt64Bytes);
-    assertEquals(Array.from(bigInt64Result), Array.from(bigInt64Test), "BigInt64Array auto-tag failed");
-    
-    const float32Test = new Float32Array([1.5, -2.5, 3.5]);
-    const float32Bytes = zserialize(float32Test);
-    const float32Result = zdeserialize(ZD.float32array(), float32Bytes);
-    assertEquals(Array.from(float32Result), Array.from(float32Test), "Float32Array auto-tag failed");
-    
-    const float64Test = new Float64Array([1.1234567890123456, -2.1234567890123456]);
-    const float64Bytes = zserialize(float64Test);
-    const float64Result = zdeserialize(ZD.float64array(), float64Bytes);
-    assertEquals(Array.from(float64Result), Array.from(float64Test), "Float64Array auto-tag failed");
-});
-
-Deno.test("Serialization - Array and Map Edge Cases for Branch Coverage", () => {
-    // Test array with single element to ensure getDefaultSerializationTag is called
-    const singleElementArray = [42];
-    const singleBytes = zserialize(singleElementArray);
-    const singleResult = zdeserialize(ZD.array(ZD.number()), singleBytes);
-    assertEquals(singleResult, singleElementArray, "Single element array failed");
-    
-    // Test array with multiple elements of same type
-    const multiElementArray = [1, 2, 3, 4, 5];
-    const multiBytes = zserialize(multiElementArray);
-    const multiResult = zdeserialize(ZD.array(ZD.number()), multiBytes);
-    assertEquals(multiResult, multiElementArray, "Multi element array failed");
-    
-    // Test Map with single entry to trigger the map branch logic
-    const singleEntryMap = new Map<string, boolean>();
-    singleEntryMap.set("test", true);
-    const mapBytes = zserialize(singleEntryMap);
-    const mapResult = zdeserialize(ZD.map(ZD.string(), ZD.boolean()), mapBytes);
-    assertEquals(Array.from(mapResult.entries()), Array.from(singleEntryMap.entries()), "Single entry map failed");
-    
-    // Test Map with multiple entries
-    const multiEntryMap = new Map<number, string>();
-    multiEntryMap.set(1, "one");
-    multiEntryMap.set(2, "two");
-    multiEntryMap.set(3, "three");
-    const multiMapBytes = zserialize(multiEntryMap);
-    const multiMapResult = zdeserialize(ZD.map(ZD.number(), ZD.string()), multiMapBytes);
-    assertEquals(Array.from(multiMapResult.entries()), Array.from(multiEntryMap.entries()), "Multi entry map failed");
-    
-    // Test edge case: Map where entries().next() might have edge behavior
-    const complexMap = new Map<bigint, Uint8Array>();
-    complexMap.set(123n, new Uint8Array([1, 2, 3]));
-    const complexMapBytes = zserialize(complexMap);
-    const complexMapResult = zdeserialize(ZD.map(ZD.bigint(), ZD.uint8array()), complexMapBytes);
-    assertEquals(complexMapResult.size, complexMap.size, "Complex map size mismatch");
-    
-    // Verify the content
-    for (const [key, value] of complexMap.entries()) {
-        const resultValue = complexMapResult.get(key);
-        assert(resultValue !== undefined, "Map key not found in result");
-        assertEquals(Array.from(resultValue), Array.from(value), "Map value mismatch");
-    }
-});
-
-Deno.test("Serialization - Edge Cases for Serialization Methods to Improve Coverage", () => {
-    // Test direct array serialization with no type specified
-    const serializer = new ZBytesSerializer();
-    const testArray = [10, 20, 30];
-    serializer.serializeArray(testArray); // This should trigger getDefaultSerializationTag
-    const arrayBytes = serializer.finish();
-    
-    const deserializer = new ZBytesDeserializer(arrayBytes);
-    const arrayResult = deserializer.deserializeArray(ZD.number());
-    assertEquals(arrayResult, testArray, "Direct array serialization failed");
-    
-    // Test direct map serialization with no types specified
-    const mapSerializer = new ZBytesSerializer();
-    const testMap = new Map<string, number>();
-    testMap.set("a", 1);
-    testMap.set("b", 2);
-    mapSerializer.serializeMap(testMap); // This should trigger getDefaultSerializationTag for both key and value
-    const mapBytes = mapSerializer.finish();
-    
-    const mapDeserializer = new ZBytesDeserializer(mapBytes);
-    const mapResult = mapDeserializer.deserializeMap(ZD.string(), ZD.number());
-    assertEquals(Array.from(mapResult.entries()), Array.from(testMap.entries()), "Direct map serialization failed");
-    
-    // Test empty array case where array.length == 0 (should have t=undefined)
-    const emptyArraySerializer = new ZBytesSerializer();
-    const emptyArray: number[] = [];
-    emptyArraySerializer.serializeArray(emptyArray);
-    const emptyArrayBytes = emptyArraySerializer.finish();
-    
-    const emptyArrayDeserializer = new ZBytesDeserializer(emptyArrayBytes);
-    const emptyArrayResult = emptyArrayDeserializer.deserializeArray(ZD.number());
-    assertEquals(emptyArrayResult, emptyArray, "Empty array direct serialization failed");
-    
-    // Test empty map case where map.size == 0
-    const emptyMapSerializer = new ZBytesSerializer();
-    const emptyMap = new Map<string, number>();
-    emptyMapSerializer.serializeMap(emptyMap);
-    const emptyMapBytes = emptyMapSerializer.finish();
-    
-    const emptyMapDeserializer = new ZBytesDeserializer(emptyMapBytes);
-    const emptyMapResult = emptyMapDeserializer.deserializeMap(ZD.string(), ZD.number());
-    assertEquals(Array.from(emptyMapResult.entries()), Array.from(emptyMap.entries()), "Empty map direct serialization failed");
-});
-
-Deno.test("Serialization - Testing Big Endian Path Coverage", () => {
-    // Create test data for all typed arrays to ensure both little and big endian paths are covered
-    // Even though we can't change the system endianness, we want to ensure all serialization calls are made
-    
-    const uint16Data = new Uint16Array([0x1234, 0x5678]);
-    const uint16Bytes = zserialize(uint16Data, ZS.uint16array());
-    const uint16Result = zdeserialize(ZD.uint16array(), uint16Bytes);
-    assertEquals(Array.from(uint16Result), Array.from(uint16Data), "Uint16Array serialization path failed");
-    
-    const uint32Data = new Uint32Array([0x12345678, 0x9ABCDEF0]);
-    const uint32Bytes = zserialize(uint32Data, ZS.uint32array());
-    const uint32Result = zdeserialize(ZD.uint32array(), uint32Bytes);
-    assertEquals(Array.from(uint32Result), Array.from(uint32Data), "Uint32Array serialization path failed");
-    
-    const bigUint64Data = new BigUint64Array([0x123456789ABCDEFn, 0xFEDCBA9876543210n]);
-    const bigUint64Bytes = zserialize(bigUint64Data, ZS.biguint64array());
-    const bigUint64Result = zdeserialize(ZD.biguint64array(), bigUint64Bytes);
-    assertEquals(Array.from(bigUint64Result), Array.from(bigUint64Data), "BigUint64Array serialization path failed");
-    
-    const int16Data = new Int16Array([-0x1234, 0x5678]);
-    const int16Bytes = zserialize(int16Data, ZS.int16array());
-    const int16Result = zdeserialize(ZD.int16array(), int16Bytes);
-    assertEquals(Array.from(int16Result), Array.from(int16Data), "Int16Array serialization path failed");
-    
-    const int32Data = new Int32Array([-0x12345678, 0x7ABCDEF0]);
-    const int32Bytes = zserialize(int32Data, ZS.int32array());
-    const int32Result = zdeserialize(ZD.int32array(), int32Bytes);
-    assertEquals(Array.from(int32Result), Array.from(int32Data), "Int32Array serialization path failed");
-    
-    const bigInt64Data = new BigInt64Array([-0x123456789ABCDEFn, 0x7EDCBA9876543210n]);
-    const bigInt64Bytes = zserialize(bigInt64Data, ZS.bigint64array());
-    const bigInt64Result = zdeserialize(ZD.bigint64array(), bigInt64Bytes);
-    assertEquals(Array.from(bigInt64Result), Array.from(bigInt64Data), "BigInt64Array serialization path failed");
-    
-    const float32Data = new Float32Array([3.14159, -2.71828, 1.41421, -0.57722]);
-    const float32Bytes = zserialize(float32Data, ZS.float32array());
-    const float32Result = zdeserialize(ZD.float32array(), float32Bytes);
-    assertEquals(Array.from(float32Result), Array.from(float32Data), "Float32Array serialization path failed");
-    
-    const float64Data = new Float64Array([Math.PI, -Math.E, Math.SQRT2, -Math.LN2]);
-    const float64Bytes = zserialize(float64Data, ZS.float64array());
-    const float64Result = zdeserialize(ZD.float64array(), float64Bytes);
-    assertEquals(Array.from(float64Result), Array.from(float64Data), "Float64Array serialization path failed");
-});
-
-Deno.test("Serialization - Individual Number Format Testing for Complete Coverage", () => {
-    // Test all number formats individually to ensure complete coverage
-    const testValues = {
-        float64: [1.7976931348623157e+308, -1.7976931348623157e+308, 0.0, -0.0],
-        float32: [3.4028235e+38, -3.4028235e+38, 1.175494e-38, -1.175494e-38],
-        int64: [9007199254740991, -9007199254740991, 0, 1, -1],
-        uint64: [9007199254740991, 0, 1, 2],
-        int32: [2147483647, -2147483648, 0, 1, -1],
-        uint32: [4294967295, 0, 1, 2],
-        int16: [32767, -32768, 0, 1, -1],
-        uint16: [65535, 0, 1, 2],
-        int8: [127, -128, 0, 1, -1],
-        uint8: [255, 0, 1, 2]
-    };
-    
-    // Test Float64
-    for (const val of testValues.float64) {
-        const bytes = zserialize(val, ZS.number(NumberFormat.Float64));
-        const result = zdeserialize(ZD.number(NumberFormat.Float64), bytes);
-        assertEquals(result, val, `Float64 failed for ${val}`);
-    }
-    
-    // Test Float32
-    for (const val of testValues.float32) {
-        const bytes = zserialize(val, ZS.number(NumberFormat.Float32));
-        const result = zdeserialize(ZD.number(NumberFormat.Float32), bytes);
-        // Use appropriate tolerance for Float32 precision
-        assert(Math.abs(result - val) < Math.abs(val) * 1e-6 || Math.abs(result - val) < 1e-30, `Float32 failed for ${val}, got ${result}`);
-    }
-    
-    // Test Int64
-    for (const val of testValues.int64) {
-        const bytes = zserialize(val, ZS.number(NumberFormat.Int64));
-        const result = zdeserialize(ZD.number(NumberFormat.Int64), bytes);
-        assertEquals(result, val, `Int64 failed for ${val}`);
-    }
-    
-    // Test Uint64
-    for (const val of testValues.uint64) {
-        const bytes = zserialize(val, ZS.number(NumberFormat.Uint64));
-        const result = zdeserialize(ZD.number(NumberFormat.Uint64), bytes);
-        assertEquals(result, val, `Uint64 failed for ${val}`);
-    }
-    
-    // Test Int32
-    for (const val of testValues.int32) {
-        const bytes = zserialize(val, ZS.number(NumberFormat.Int32));
-        const result = zdeserialize(ZD.number(NumberFormat.Int32), bytes);
-        assertEquals(result, val, `Int32 failed for ${val}`);
-    }
-    
-    // Test Uint32
-    for (const val of testValues.uint32) {
-        const bytes = zserialize(val, ZS.number(NumberFormat.Uint32));
-        const result = zdeserialize(ZD.number(NumberFormat.Uint32), bytes);
-        assertEquals(result, val, `Uint32 failed for ${val}`);
-    }
-    
-    // Test Int16
-    for (const val of testValues.int16) {
-        const bytes = zserialize(val, ZS.number(NumberFormat.Int16));
-        const result = zdeserialize(ZD.number(NumberFormat.Int16), bytes);
-        assertEquals(result, val, `Int16 failed for ${val}`);
-    }
-    
-    // Test Uint16
-    for (const val of testValues.uint16) {
-        const bytes = zserialize(val, ZS.number(NumberFormat.Uint16));
-        const result = zdeserialize(ZD.number(NumberFormat.Uint16), bytes);
-        assertEquals(result, val, `Uint16 failed for ${val}`);
-    }
-    
-    // Test Int8
-    for (const val of testValues.int8) {
-        const bytes = zserialize(val, ZS.number(NumberFormat.Int8));
-        const result = zdeserialize(ZD.number(NumberFormat.Int8), bytes);
-        assertEquals(result, val, `Int8 failed for ${val}`);
-    }
-    
-    // Test Uint8
-    for (const val of testValues.uint8) {
-        const bytes = zserialize(val, ZS.number(NumberFormat.Uint8));
-        const result = zdeserialize(ZD.number(NumberFormat.Uint8), bytes);
-        assertEquals(result, val, `Uint8 failed for ${val}`);
-    }
-});
-
-Deno.test("Serialization - BigInt Format Testing for Complete Coverage", () => {
-    // Test both BigInt formats with edge values
-    const bigintValues = [
-        0n, 1n, -1n, 
-        9223372036854775807n, -9223372036854775808n, // Int64 limits
-        18446744073709551615n, // Uint64 max
-        123456789012345678n, -987654321098765432n
-    ];
-    
-    // Test BigIntFormat.Int64
-    for (const val of bigintValues) {
-        try {
-            const bytes = zserialize(val, ZS.bigint(BigIntFormat.Int64));
-            const result = zdeserialize(ZD.bigint(BigIntFormat.Int64), bytes);
-            assertEquals(result, val, `BigInt Int64 failed for ${val}`);
-        } catch (e) {
-            // Some values might be out of range for Int64, that's expected
-            if (val > 9223372036854775807n || val < -9223372036854775808n) {
-                // Expected for out of range values
-                continue;
-            } else {
-                throw e;
-            }
-        }
-    }
-    
-    // Test BigIntFormat.Uint64 (only positive values)
-    const uint64Values = bigintValues.filter(v => v >= 0n);
-    for (const val of uint64Values) {
-        try {
-            const bytes = zserialize(val, ZS.bigint(BigIntFormat.Uint64));
-            const result = zdeserialize(ZD.bigint(BigIntFormat.Uint64), bytes);
-            assertEquals(result, val, `BigInt Uint64 failed for ${val}`);
-        } catch (e) {
-            // Some values might be out of range for Uint64, that's expected
-            if (val > 18446744073709551615n || val < 0n) {
-                // Expected for out of range values
-                continue;
-            } else {
-                throw e;
-            }
-        }
-    }
-});
